pyproject.toml
[project]
name = "ezsn-notify-email"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi[standard]>=0.135.3",
    "python-dotenv>=1.2.2",
]


[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
fix = true

[dependency-groups]
dev = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.9",
]
